refactor: App.tsx の責務分割と application/hooks への状態抽出による再構築
