+++ Frontend/src/main.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './styles/index.css'
-import App from './app/App'
+import App from './App'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
